Llms: move out icons
