+++ src/common/components/icons/AnthropicIcon.tsx
