+++ src/common/components/icons/AzureIcon.tsx
