+++ src/common/components/icons/OobaboogaIcon.tsx
