+++ src/common/components/icons/OpenAIIcon.tsx
