+++ src/modules/llms/anthropic/anthropic.vendor.ts
@@ -1,18 +1,19 @@
 import { apiAsync } from '~/modules/trpc/trpc.client';
 
+import { AnthropicIcon } from '~/common/components/icons/AnthropicIcon';
+
 import { DLLM, ModelVendor } from '../llm.types';
 import { VChatMessageIn, VChatMessageOut } from '../llm.client';
 
-import { LLMOptionsOpenAI } from '~/modules/llms/openai/openai.vendor';
-import { OpenAILLMOptions } from '~/modules/llms/openai/OpenAILLMOptions';
+import { LLMOptionsOpenAI } from '../openai/openai.vendor';
+import { OpenAILLMOptions } from '../openai/OpenAILLMOptions';
 
-import { AnthropicIcon } from './AnthropicIcon';
 import { AnthropicSourceSetup } from './AnthropicSourceSetup';
+
 
 // special symbols
 export const hasServerKeyAnthropic = !!process.env.HAS_SERVER_KEY_ANTHROPIC;
 export const isValidAnthropicApiKey = (apiKey?: string) => !!apiKey && apiKey.startsWith('sk-') && apiKey.length > 40;
-
 
 export interface SourceSetupAnthropic {
   anthropicKey: string;
