+++ src/modules/llms/openrouter/openrouter.vendor.tsx
@@ -1,10 +1,12 @@
+import { OpenRouterIcon } from '~/common/components/icons/OpenRouterIcon';
+
 import { ModelVendor } from '../llm.types';
 
-import { LLMOptionsOpenAI, ModelVendorOpenAI } from '~/modules/llms/openai/openai.vendor';
-import { OpenAILLMOptions } from '~/modules/llms/openai/OpenAILLMOptions';
+import { LLMOptionsOpenAI, ModelVendorOpenAI } from '../openai/openai.vendor';
+import { OpenAILLMOptions } from '../openai/OpenAILLMOptions';
 
-import { OpenRouterIcon } from './OpenRouterIcon';
 import { OpenRouterSourceSetup } from './OpenRouterSourceSetup';
+
 
 // special symbols
 export const isValidOpenRouterKey = (apiKey?: string) => !!apiKey && apiKey.startsWith('sk-or-') && apiKey.length > 40;
@@ -15,7 +17,6 @@ export interface SourceSetupOpenRouter {
   oaiHost: string;
 }
 
-
 /**
  * NOTE: the support is just started and incomplete - in particular it depends on some code that
  * hasn't been merged yet.
@@ -23,7 +24,7 @@ export interface SourceSetupOpenRouter {
  * Completion:
  *  [x] raise instanceLimit from 0 to 1 to continue development
  *  [x] add support to the OpenAI Router and Streaming function to add the headers required by OpenRouter (done in the access function)
- *  [ ] merge the server-side models remapping from Azure OpenAI - not needed, using client-side remapping for now
+ *  [~] merge the server-side models remapping from Azure OpenAI - not needed, using client-side remapping for now
  *  [x] decide whether to do UI work to improve the appearance - prioritized models
  *  [x] works!
  */
